# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QueueCare AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PatientStatus = typeof PatientStatus[keyof typeof PatientStatus];


export const PatientStatus = {
  waiting: 'waiting',
  called: 'called',
  in_consultation: 'in_consultation',
  completed: 'completed',
  skipped: 'skipped',
} as const;

export interface Patient {
  id: number;
  tokenNumber: number;
  name: string;
  phone: string;
  visitType: string;
  /** @nullable */
  symptoms?: string | null;
  status: PatientStatus;
  /** @nullable */
  estimatedWaitMinutes?: number | null;
  /** @nullable */
  patientsAhead?: number | null;
  createdAt: string;
  /** @nullable */
  calledAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
}

export interface PatientInput {
  name: string;
  phone: string;
  visitType: string;
  symptoms?: string;
}

export type PatientTrackingInfoQueueStatus = typeof PatientTrackingInfoQueueStatus[keyof typeof PatientTrackingInfoQueueStatus];


export const PatientTrackingInfoQueueStatus = {
  active: 'active',
  paused: 'paused',
  closed: 'closed',
} as const;

export interface PatientTrackingInfo {
  patient: Patient;
  /** @nullable */
  currentServingToken: number | null;
  patientsAhead: number;
  estimatedWaitMinutes: number;
  queueStatus: PatientTrackingInfoQueueStatus;
}

export type QueueStateQueueStatus = typeof QueueStateQueueStatus[keyof typeof QueueStateQueueStatus];


export const QueueStateQueueStatus = {
  active: 'active',
  paused: 'paused',
  closed: 'closed',
} as const;

export interface QueueState {
  patients: Patient[];
  /** @nullable */
  currentToken: number | null;
  totalWaiting: number;
  totalCompleted: number;
  averageWaitMinutes: number;
  queueStatus: QueueStateQueueStatus;
  /** @nullable */
  estimatedQueueFinish?: string | null;
}

export interface Doctor {
  id: number;
  name: string;
  specialization: string;
  isActive: boolean;
  /** @nullable */
  currentPatientId?: number | null;
}

export interface DoctorView {
  doctor: Doctor;
  currentPatient?: Patient;
  upcomingPatients: Patient[];
}

export interface AnalyticsSummary {
  patientsToday: number;
  patientsServed: number;
  averageWaitMinutes: number;
  averageConsultationMinutes: number;
  peakHour: string;
  queueEfficiency: number;
  totalWaiting?: number;
}

export interface HourlyData {
  hour: string;
  patients: number;
  avgWait: number;
}

export type AiInsightsQueueHealth = typeof AiInsightsQueueHealth[keyof typeof AiInsightsQueueHealth];


export const AiInsightsQueueHealth = {
  excellent: 'excellent',
  good: 'good',
  moderate: 'moderate',
  critical: 'critical',
} as const;

export interface AiInsights {
  queueHealth: AiInsightsQueueHealth;
  predictedFinishTime: string;
  avgConsultationMinutes: number;
  recommendations: string[];
  currentLoad?: number;
}

export interface ClinicSettings {
  id: number;
  clinicName: string;
  doctorName: string;
  specialization?: string;
  maxQueueSize: number;
  avgConsultationMinutes: number;
  openTime: string;
  closeTime: string;
  notificationsEnabled: boolean;
  smsEnabled?: boolean;
  autoCallNext?: boolean;
}

export interface ClinicSettingsUpdate {
  clinicName?: string;
  doctorName?: string;
  specialization?: string;
  maxQueueSize?: number;
  avgConsultationMinutes?: number;
  openTime?: string;
  closeTime?: string;
  notificationsEnabled?: boolean;
  smsEnabled?: boolean;
  autoCallNext?: boolean;
}