# lib/api-zod/src/generated/types/aiInsights.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QueueCare AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiInsightsQueueHealth } from './aiInsightsQueueHealth';

export interface AiInsights {
  queueHealth: AiInsightsQueueHealth;
  predictedFinishTime: string;
  avgConsultationMinutes: number;
  recommendations: string[];
  currentLoad?: number;
}