# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QueueCare AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsSummary {
  patientsToday: number;
  patientsServed: number;
  averageWaitMinutes: number;
  averageConsultationMinutes: number;
  peakHour: string;
  queueEfficiency: number;
  totalWaiting?: number;
}