# lib/api-zod/src/generated/types/clinicSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QueueCare AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClinicSettings {
  id: number;
  clinicName: string;
  doctorName: string;
  specialization?: string;
  maxQueueSize: number;
  avgConsultationMinutes: number;
  openTime: string;
  closeTime: string;
  notificationsEnabled: boolean;
  smsEnabled?: boolean;
  autoCallNext?: boolean;
}