# lib/api-zod/src/generated/types/doctorView.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QueueCare AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Doctor } from './doctor';
import type { Patient } from './patient';

export interface DoctorView {
  doctor: Doctor;
  currentPatient?: Patient;
  upcomingPatients: Patient[];
}