# lib/api-zod/src/generated/types/hourlyData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QueueCare AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HourlyData {
  hour: string;
  patients: number;
  avgWait: number;
}