# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QueueCare AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PatientStatus } from './patientStatus';

export interface Patient {
  id: number;
  tokenNumber: number;
  name: string;
  phone: string;
  visitType: string;
  /** @nullable */
  symptoms?: string | null;
  status: PatientStatus;
  /** @nullable */
  estimatedWaitMinutes?: number | null;
  /** @nullable */
  patientsAhead?: number | null;
  createdAt: string;
  /** @nullable */
  calledAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
}